+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
-- [...]
+- **[BREAKING CHANGE]** Rename `WarningModal` to `ConfirmationModal`
+- **[BREAKING CHANGE]** `ConfirmationModal` remove `title` and `displayCloseButton` props
+- **[BREAKING CHANGE]** `ConfirmationModal` add `status` prop to display it as `WARNING/REMINDER`
 
 # v0.19.1 (19/12/2018)